Add option to auto reload the geo ip db file
This version of the GeoIp module was modified to cache the database file
in memory. This changeset adds a check for the last modified time of the
file and forces the reload if necessary. This check can be enabled and
disabled in php.ini with the "geoip.autoreload_db" option set to 1
(defaults to 0).

It also provides a new function to force the reload of the file
explictly, geip_db_reload().